semana-3/Modulo-2/ex02/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
chardet==5.2.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
ruff==0.11.9
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
